stupid junit

+++ test/net/todd/bible/scripturelookup/server/BibleDaoIntegrationTest.java
@@ -48,18 +48,18 @@ public class BibleDaoIntegrationTest {
 				1, 2, 3));
 		assertEquals(3, john1.size());
 		assertEquals("John", john1.get(0).getBook());
-		assertEquals(1, john1.get(0).getChapter());
-		assertEquals(1, john1.get(0).getVerse());
+		assertEquals(1, john1.get(0).getChapter().intValue());
+		assertEquals(1, john1.get(0).getVerse().intValue());
 		assertEquals(
 				"In the beginning was the Word, and the Word was with God, and the Word was God.",
 				john1.get(0).getText());
 		assertEquals("John", john1.get(1).getBook());
-		assertEquals(1, john1.get(1).getChapter());
-		assertEquals(2, john1.get(1).getVerse());
+		assertEquals(1, john1.get(1).getChapter().intValue());
+		assertEquals(2, john1.get(1).getVerse().intValue());
 		assertEquals("He was with God in the beginning", john1.get(1).getText());
 		assertEquals("John", john1.get(2).getBook());
-		assertEquals(1, john1.get(2).getChapter());
-		assertEquals(3, john1.get(2).getVerse());
+		assertEquals(1, john1.get(2).getChapter().intValue());
+		assertEquals(3, john1.get(2).getVerse().intValue());
 		assertEquals(
 				"Through him all things were made; without him nothing was made that has been made.",
 				john1.get(2).getText());
